XMLHttpRequest/responsetext-decoding.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: responseText decoding</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-responsetext-attribute" data-tested-assertations="following::ol[1]/li[4]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#text-response-entity-body" data-tested-assertations="following::ol[1]/li[2] following::ol[1]/li[3] following::ol[1]/li[4] following::ol[1]/li[5]" />
  </head>
  <body>
    <script>
      function html(content) {
        return "<!doctype html><meta charset=windows-1252><x>" + content + "</x>";
      }
      function encoded_html(encoded_content) {
        return encodeURIComponent("<!doctype html><meta charset=windows-1252><x>") + encoded_content + encodeURIComponent("<\/x>");
      }
      function xml(content) {
        return "<?xml version='1.0' encoding='windows-1252'?><x>" + content + "</x>";
      }
      function encoded_xml(encoded_content) {
        return encodeURIComponent("<?xml version='1.0' encoding='windows-1252'?><x>") + encoded_content + encodeURIComponent("<\/x>");
      }
      function request(type, input, output, responseType) {
        var test = async_test(document.title + " (" + type + " " + input + " " + (responseType ? " " + responseType : "default") + ")");
        test.step(function() {
          var client = new XMLHttpRequest();
          if (responseType !== undefined) {
            client.responseType = responseType;
          }
          client.open("GET", "resources/status.py?content=" + input + "&type=" + encodeURIComponent(type), true);
          client.onload = test.step_func_done(() => {
            assert_equals(client.responseText, output);
          })
          client.send(null);
        })
      }

      // "default" response type
      // An XML-ish response is sniffed.
      request("application/xml", encoded_xml("%FF"), xml("\u00FF"));
      // An HTML-ish response isn't sniffed.
      request("text/html", encoded_html("%FF"), html("\uFFFD"));
      request("application/xml;charset=utf-8", encoded_xml("%FF"), xml("\uFFFD"));
      request("application/xml;charset=windows-1252", encoded_xml("%FF"), xml("\u00FF"));
      request("text/html;charset=utf-8", encoded_html("%FF"), html("\uFFFD"));
      request("text/html;charset=windows-1252", encoded_html("%FF"), html("\u00FF"));
      request("text/plain;charset=windows-1252", "%FF", "\u00FF");
      request("text/plain", "%FF", "\uFFFD");
      request("text/plain", "%FE%FF", "");
      request("text/plain", "%FE%FF%FE%FF", "\uFEFF");
      request("text/plain", "%EF%BB%BF", "");
      request("text/plain", "%EF%BB%BF%EF%BB%BF", "\uFEFF");
      request("text/plain", "%C2", "\uFFFD");
      request("text/xml", "%FE%FF", "");
      request("text/xml", "%FE%FF%FE%FF", "\uFEFF");
      request("text/xml", "%EF%BB%BF", "");
      request("text/xml", "%EF%BB%BF%EF%BB%BF", "\uFEFF");
      request("text/plain", "%E3%81%B2", "\u3072");

      // "text" response type
      // An XML-ish response isn't sniffed.
      request("application/xml", encoded_xml("%FF"), xml("\uFFFD"), "text");
      // An HTML-ish response isn't sniffed.
      request("text/html", encoded_html("%FF"), html("\uFFFD"), "text");
      request("text/plain;charset=windows-1252", "%FF", "\u00FF", "text");
      request("text/plain", "%FF", "\uFFFD", "text");
      request("text/plain", "%FE%FF", "", "text");
      request("text/plain", "%FE%FF%FE%FF", "\uFEFF", "text");
      request("text/plain", "%EF%BB%BF", "", "text");
      request("text/plain", "%EF%BB%BF%EF%BB%BF", "\uFEFF", "text");
      request("text/plain", "%C2", "\uFFFD", "text");
      request("text/xml", "%FE%FF", "", "text");
      request("text/xml", "%FE%FF%FE%FF", "\uFEFF", "text");
      request("text/xml", "%EF%BB%BF", "", "text");
      request("text/xml", "%EF%BB%BF%EF%BB%BF", "\uFEFF", "text");
      request("text/plain", "%E3%81%B2", "\u3072", "text");
    </script>
  </body>
</html>
